requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
